Compacted migrations, added new fields, renamed image sources
Added JSON data to Post
Added UUID to PostImage
Renamed image author/owner to source in Post and PostImage

+++ db/migrate/20180706000000_create_editorial_members.rb
@@ -1,21 +1,22 @@
+# frozen_string_literal: true
+
+# Table for editorial members
 class CreateEditorialMembers < ActiveRecord::Migration[5.2]
   def up
-    unless EditorialMember.table_exists?
-      create_table :editorial_members do |t|
-        t.timestamps
-        t.references :user, null: false, foreign_key: { on_update: :cascade, on_delete: :cascade }
-        t.boolean :visible, default: true, null: false
-        t.integer :priority, limit: 2, default: 1, null: false
-        t.string :title
-        t.text :lead
-        t.text :about
-      end
+    return if EditorialMember.table_exists?
+
+    create_table :editorial_members, comment: 'Editorial member' do |t|
+      t.references :user, null: false, foreign_key: { on_update: :cascade, on_delete: :cascade }
+      t.timestamps
+      t.boolean :visible, default: true, null: false
+      t.integer :priority, limit: 2, default: 1, null: false
+      t.string :title
+      t.text :lead
+      t.text :about
     end
   end
 
   def down
-    if EditorialMember.table_exists?
-      drop_table :editorial_members
-    end
+    drop_table :editorial_members if EditorialMember.table_exists?
   end
 end
